# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.ErrorResponse:
    properties:
      code:
        type: integer
      details:
        type: string
      message:
        type: string
    type: object
  dto.FiscalConfigListResponse:
    properties:
      configurations:
        items:
          $ref: '#/definitions/dto.FiscalConfigResponse'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  dto.FiscalConfigRequest:
    properties:
      branch_id:
        type: string
      certificate_id:
        type: string
      contingency_enabled:
        type: boolean
      fiscal_csc:
        description: Configurações Gerais
        type: string
      fiscal_csc_id:
        type: string
      nfce_csc_id:
        type: string
      nfce_csc_token:
        type: string
      nfce_environment:
        $ref: '#/definitions/fiscal.FiscalEnvironment'
      nfce_next_number:
        minimum: 1
        type: integer
      nfce_series:
        description: Configurações NFCe
        type: string
      nfe_csc_id:
        type: string
      nfe_csc_token:
        type: string
      nfe_environment:
        $ref: '#/definitions/fiscal.FiscalEnvironment'
      nfe_next_number:
        minimum: 1
        type: integer
      nfe_series:
        description: Configurações NFe
        type: string
      print_danfe_mode:
        allOf:
        - $ref: '#/definitions/fiscal.PrintMode'
        description: Configurações de impressão
      printer_name:
        type: string
      printer_paper_size:
        type: string
      smtp_host:
        description: SMTP para envio de documentos fiscais
        type: string
      smtp_password:
        type: string
      smtp_port:
        type: integer
      smtp_username:
        type: string
    required:
    - branch_id
    - nfce_environment
    - nfce_next_number
    - nfce_series
    - nfe_environment
    - nfe_next_number
    - nfe_series
    type: object
  dto.FiscalConfigResponse:
    properties:
      branch_id:
        type: string
      branch_name:
        type: string
      certificate_id:
        type: string
      certificate_name:
        type: string
      contingency_enabled:
        type: boolean
      created_at:
        type: string
      fiscal_csc:
        description: Configurações Gerais
        type: string
      fiscal_csc_id:
        type: string
      id:
        type: string
      nfce_csc_id:
        type: string
      nfce_csc_token:
        type: string
      nfce_environment:
        $ref: '#/definitions/fiscal.FiscalEnvironment'
      nfce_next_number:
        type: integer
      nfce_series:
        description: Configurações NFCe
        type: string
      nfe_csc_id:
        type: string
      nfe_csc_token:
        type: string
      nfe_environment:
        $ref: '#/definitions/fiscal.FiscalEnvironment'
      nfe_next_number:
        type: integer
      nfe_series:
        description: Configurações NFe
        type: string
      print_danfe_mode:
        allOf:
        - $ref: '#/definitions/fiscal.PrintMode'
        description: Configurações de impressão
      printer_name:
        type: string
      printer_paper_size:
        type: string
      smtp_host:
        description: SMTP para envio de documentos fiscais
        type: string
      smtp_port:
        type: integer
      smtp_username:
        type: string
      updated_at:
        type: string
    type: object
...
  dto.SuccessResponse:
    properties:
      data: {}
      message:
        type: string
    type: object
...
  fiscal.FiscalEnvironment:
    enum:
    - production
    - homologation
    type: string
    x-enum-varnames:
    - Production
    - Homologation
  fiscal.PrintMode:
    enum:
    - normal
    - contingency
    - none
    type: string
    x-enum-varnames:
    - Normal
    - Contingency
    - None
host: localhost:8084
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API para o sistema ERP de supermercados
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: SuperERP API
  version: "1.0"
paths:
...
  /fiscal/configs:
    get:
      description: Lista as configurações fiscais com paginação
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 'Número da página (padrão: 1)'
        in: query
        name: page
        type: integer
      - description: 'Tamanho da página (padrão: 10)'
        in: query
        name: page_size
        type: integer
      - description: Filtrar por filial
        in: query
        name: branch_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.FiscalConfigListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Listar configurações fiscais
      tags:
      - Configurações Fiscais
    post:
      consumes:
      - application/json
      description: Cria uma nova configuração fiscal
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Dados da configuração fiscal
        in: body
        name: config
        required: true
        schema:
          $ref: '#/definitions/dto.FiscalConfigRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.FiscalConfigResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Criar configuração fiscal
      tags:
      - Configurações Fiscais
  /fiscal/configs/{id}:
    delete:
      description: Remove uma configuração fiscal pelo ID
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID da configuração fiscal
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Excluir configuração fiscal
      tags:
      - Configurações Fiscais
    get:
      description: Busca uma configuração fiscal pelo ID
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID da configuração fiscal
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.FiscalConfigResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Obter configuração fiscal
      tags:
      - Configurações Fiscais
    put:
      consumes:
      - application/json
      description: Atualiza uma configuração fiscal existente
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID da configuração fiscal
        in: path
        name: id
        required: true
        type: string
      - description: Dados da configuração fiscal
        in: body
        name: config
        required: true
        schema:
          $ref: '#/definitions/dto.FiscalConfigRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.FiscalConfigResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Atualizar configuração fiscal
      tags:
      - Configurações Fiscais
  /fiscal/configs/branch/{branch_id}:
    get:
      description: Busca a configuração fiscal para uma filial específica
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID da filial
        in: path
        name: branch_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.FiscalConfigResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Obter configuração fiscal por filial
      tags:
      - Configurações Fiscais
  /fiscal/configs/branch/{branch_id}/contingency:
    post:
      consumes:
      - application/json
      description: Ativa ou desativa o modo de contingência para a filial
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID da filial
        in: path
        name: branch_id
        required: true
        type: string
      - description: Dados de contingência
        in: body
        name: contingency
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Atualizar modo de contingência
      tags:
      - Configurações Fiscais
  /fiscal/configs/branch/{branch_id}/increment-nfce:
    post:
      description: Incrementa o próximo número de NFCe para a filial
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID da filial
        in: path
        name: branch_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Incrementar numeração de NFCe
      tags:
      - Configurações Fiscais
  /fiscal/configs/branch/{branch_id}/increment-nfe:
    post:
      description: Incrementa o próximo número de NFe para a filial
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID da filial
        in: path
        name: branch_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Incrementar numeração de NFe
      tags:
      - Configurações Fiscais
...
securityDefinitions:
  Bearer:
    description: 'Cabeçalho de autenticação JWT usando o esquema Bearer. Exemplo:
      "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"